Rename defaultSettings.js to default-settings.js
