+++ Source/javascripts/constants/default-settings.js
